Create framework for stepper implementation
